put utils in separate file and add tests
